Refactor and simplify package structure
