--- .editorconfig
@@ -1,12 +1,0 @@
-; Please use these settings, not your own
-root = true
-
-; Unix-style newlines, and small indenting
-[*]
-end_of_line = lf
-insert_final_newline = true
-indent_style = space
-indent_size = 2
-trim_trailing_whitespace = true
-charset = utf-8
-
