+++ .gitignore
@@ -5,7 +5,3 @@ node_modules/
 # Playground / Testing Files #
 ##############################
 index.html
-
-# Build files #
-###############
-dist/
