+++ .jshintrc
@@ -1,47 +1,20 @@
 {
-  "predef": [
-    "TeoriaNote",
-    "TeoriaInterval",
-    "TeoriaChord",
-    "TeoriaScale",
-    "teoria",
-
-    "kNotes",
-    "kNoteIndex",
-    "kDurations",
-    "kIntervals",
-    "kIntervalIndex",
-    "kQualityLong",
-    "kQualityTemp",
-    "kValidQualities",
-    "kQualityInversion",
-    "kAlterations",
-    "kChords",
-    "kChordShort",
-    "kAccidentalSign",
-    "kAccidentalValue",
-    "kStepNumber",
-    "kIntervalSolfege",
-
-    "getDistance",
-    "pad"
-  ],
-
   "maxlen":   80,
   "browser":  true,
   "devel":    true,
   "curly":    true,
   "es5":      true,
   "newcap":   true,
+  "node":     true,
   "noempty":  true,
   "nonew":    true,
   "onevar":   false,
   "plusplus": false,
   "trailing": true,
-  "eqeqeq": true,
-  "forin": true,
-  "latedef": true,
-  "undef": true,
-  "evil": false,
-  "unused": true
+  "eqeqeq":   true,
+  "forin":    true,
+  "latedef":  true,
+  "undef":    true,
+  "evil":     false,
+  "unused":   true
 }
