+++ package.json
@@ -1,20 +1,18 @@
 {
   "name": "teoria",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "description": "Music theory for JavaScript",
-  "homepage": "https://github.com/saebekassebil/teoria",
+  "homepage": "saebekassebil.github.com/teoria",
   "author": {
-    "name": "Jakob Miland <saebekassebil>",
+    "name": "Jakob Miland",
     "email": "saebekassebil@gmail.com",
     "url": "https://github.com/saebekassebil"
   },
+  "main": "teoria.js",
 
   "devDependencies": {
     "jake": ">=0.2.0",
-    "colors": ">=0.6.0",
-    "mingler": ">=0.1.3",
     "jshint": ">=0.9.0",
-    "uglify-js": "~1",
     "vows": ">= 0.6.0"
   }
 }
